requirements.txt
numpy==1.19.1
pandas==1.1.0
psycopg2==2.8.5
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.18
